test several
Former-commit-id: bac1fa8c26576fbbe4e3e84a9d626e26836305cc

+++ HFTPlatform/zHFT.InstructionBasedMarketClient.Binance/Client/BinanceInstructionBasedMarketClient.cs
@@ -199,7 +199,7 @@ namespace zHFT.InstructionBasedMarketClient.Binance.Client
                 if (LoadConfig(configFile))
                 {
 
-                    ActiveSecurities = new Dictionary<int, Security>();
+                    ActiveSecurities = new Dictionary<long, Security>();
                     ContractsTimeStamps = new Dictionary<int, DateTime>();
                     ReverseCurrency = new Dictionary<string, bool>();
